fix: rescue DoS, incorrect-shift lint, guardian trust documentation
Security fixes:
- AirAccountDelegate.initiateRescue(): block any pending-rescue override (was
  only blocking same-destination re-initiation). Once a rescue is pending, no
  guardian can override it — prevents DoS via competing initiations. EOA must
  call cancelRescue() before a new rescue can be started.

Lint fixes (incorrect-shift):
- AAStarAirAccountBase.sol: `1 << guardianIndex` → `uint256(1) << guardianIndex` (x3)
- AirAccountDelegate.sol: `uint8(1 << gIdx)` → `uint8(1) << gIdx` (x3)
  Makes type intent explicit; prevents silent truncation if index bounds grow.

Documentation:
- AirAccountDelegate.initialize(): add guardian trust warning explaining 2-of-3
  collusion can drain the account within the 2-day timelock window.

Tests (+1 net):
- Renamed test_rescue_overrideWithDifferentDestination → test_rescue_overrideBlocked_reverts
  (now correctly expects RescueAlreadyPending on any pending-rescue override)
- Added test_rescue_cancelThenReinitiate: cancel clears state, new rescue can start

391 tests, 0 failed, 0 incorrect-shift warnings.

+++ src/core/AAStarAirAccountBase.sol
@@ -952,7 +952,7 @@ abstract contract AAStarAirAccountBase {
         activeRecovery = RecoveryProposal({
             newOwner: _newOwner,
             proposedAt: block.timestamp,
-            approvalBitmap: 1 << guardianIndex,
+            approvalBitmap: uint256(1) << guardianIndex,
             cancellationBitmap: 0
         });
 
@@ -965,7 +965,7 @@ abstract contract AAStarAirAccountBase {
         if (activeRecovery.newOwner == address(0)) revert NoActiveRecovery();
 
         uint8 guardianIndex = _guardianIndex(msg.sender);
-        uint256 bit = 1 << guardianIndex;
+        uint256 bit = uint256(1) << guardianIndex;
         if (activeRecovery.approvalBitmap & bit != 0) revert AlreadyApproved();
 
         activeRecovery.approvalBitmap |= bit;
@@ -1001,7 +1001,7 @@ abstract contract AAStarAirAccountBase {
         if (activeRecovery.newOwner == address(0)) revert NoActiveRecovery();
 
         uint8 guardianIndex = _guardianIndex(msg.sender); // reverts if not guardian
-        uint256 bit = 1 << guardianIndex;
+        uint256 bit = uint256(1) << guardianIndex;
         if (activeRecovery.cancellationBitmap & bit != 0) revert AlreadyCancelVoted();
 
         activeRecovery.cancellationBitmap |= bit;
